+++ src/index.ts
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
 
-import apiGenerator from "./apiGenerator";
+import apiGenerator from "./api-generator";
 
 apiGenerator(process.argv[2]);
